fix: handling ErrNamespaceNotFound for node v0.10.0 (#953)
We have made a breaking change in node v0.10.0 regarding data requests.
https://github.com/celestiaorg/celestia-node/commit/62a0b97997b865efd76a2e8d02af8ee5d2a8e912

Previously, a "data not found" meant both "the node failed to retrieve
that data, please retry" as well as "the node retrieved the data
successfully, but your namespace is not included in the data". We have
split these errors, so now we have an `ErrNamespaceNotFound` which
indicates a successful request for empty data.

I have no idea how to test rollkit so would love some help on getting
that set up

---------

Co-authored-by: Ganesha Upadhyaya <gupadhyaya@Ganeshas-MacBook-Pro-2.local>