da: add celestia-openrpc dep (#994)
## Overview

This PR adds a dependency on `celestia-openrpc` `505ff801e818` which
will eventually replace `go-cnc`. `go-header` was pinned to `0.2.6` as
`0.2.7` is breaking tests.

Fixes #995 

## Checklist

<!-- 
Please complete the checklist to ensure that the PR is ready to be
reviewed.

IMPORTANT:
PRs should be left in Draft until the below checklist is completed.
-->

- [x] New and updated code has appropriate documentation
- [x] New and updated code has new and/or updated testing
- [x] Required CI checks are passing
- [x] Visual proof for any user facing features like CLI or
documentation updates
- [x] Linked issues closed with keywords