FileLogger for Testing (#1067)
<!--
Please read and fill out this form before submitting your PR.

Please make sure you have reviewed our contributors guide before
submitting your
first PR.
-->

## Overview

Something I have found incredibly changing when debugging tests is the
inability to use the verbose flag for doing some basic println debugging
since the stdout is flooding with the logs of the node.

To address this, I introduced a filelogger that can be used in testing
to put the logs in a temp file and keep the stdout free for debugging.

The tempfile can then still be used to investigate the logs if needed.
This will also make debugging on github action easier since we can use
the verbose flag for helpful errors and download the logfiles if needed.

I'm probably missing some updates, but I did these when trying to help
debug a test for Manav and so wanted to share.

Closes #1079 

<!-- 
Please provide an explanation of the PR, including the appropriate
context,
background, goal, and rationale. If there is an issue with this
information,
please provide a tl;dr and link the issue. 
-->

## Checklist

<!-- 
Please complete the checklist to ensure that the PR is ready to be
reviewed.

IMPORTANT:
PRs should be left in Draft until the below checklist is completed.
-->

- [ ] New and updated code has appropriate documentation
- [ ] New and updated code has new and/or updated testing
- [ ] Required CI checks are passing
- [ ] Visual proof for any user facing features like CLI or
documentation updates
- [ ] Linked issues closed with keywords

---------

Co-authored-by: Manav Aggarwal <manavaggarwal1234@gmail.com>
Co-authored-by: Tomasz Zdybał <tomek@zdybal.lap.pl>