deps: fix celestia-openrpc v0.1.2 breaking changes (#1140)
## Overview

`celestia-openrpc v0.1.2` signature for `NewDataAvailabilityHeader` now
returns an error, so this needed to be handled.

`RandEDS` and `RandShares` needed to be updated for the new `Share`
type.

## Checklist

<!-- 
Please complete the checklist to ensure that the PR is ready to be
reviewed.

IMPORTANT:
PRs should be left in Draft until the below checklist is completed.
-->

- [x] New and updated code has appropriate documentation
- [x] New and updated code has new and/or updated testing
- [x] Required CI checks are passing
- [x] Visual proof for any user facing features like CLI or
documentation updates
- [x] Linked issues closed with keywords

---------

Signed-off-by: dependabot[bot] <support@github.com>
Co-authored-by: dependabot[bot] <49699333+dependabot[bot]@users.noreply.github.com>