test: use pkg httptest to avoid port conflict (#1144)
## Overview

This PR uses pkg `httptest` to avoid port conflict when running tests
under `da/test` i.e. `TestLifecycle`, `TestRetrieve`.

This allows the tests to pass even if an address is bound to the default
port i.e. `127.0.0.1:26658`

## Checklist

<!-- 
Please complete the checklist to ensure that the PR is ready to be
reviewed.

IMPORTANT:
PRs should be left in Draft until the below checklist is completed.
-->

- [ ] New and updated code has appropriate documentation
- [ ] New and updated code has new and/or updated testing
- [ ] Required CI checks are passing
- [ ] Visual proof for any user facing features like CLI or
documentation updates
- [ ] Linked issues closed with keywords