test: update mock.Server to use httptest.Server (#1145)
## Overview

This PR improves upon #1144 by simplifying `mock.Server` to use
`httptest` directly, avoiding a duplicate `http.Server`. `Server.Start`
doesn't need a listener anymore, and it returns the URL at which the
`httptest.Server` is running.

## Checklist

<!-- 
Please complete the checklist to ensure that the PR is ready to be
reviewed.

IMPORTANT:
PRs should be left in Draft until the below checklist is completed.
-->

- [ ] New and updated code has appropriate documentation
- [ ] New and updated code has new and/or updated testing
- [ ] Required CI checks are passing
- [ ] Visual proof for any user facing features like CLI or
documentation updates
- [ ] Linked issues closed with keywords