feat: add go fmt to make fmt command (#1178)
<!--
Please read and fill out this form before submitting your PR.

Please make sure you have reviewed our contributors guide before
submitting your
first PR.
-->

## Overview

<!-- 
Please provide an explanation of the PR, including the appropriate
context,
background, goal, and rationale. If there is an issue with this
information,
please provide a tl;dr and link the issue. 
-->

Closes #223 

Instead of adding a formatted to the CI, which would introduce a diff
that would need to be committed, I extended the `make fmt` command.