Sevey/shutdown management (#1434)
<!--
Please read and fill out this form before submitting your PR.

Please make sure you have reviewed our contributors guide before
submitting your
first PR.
-->

## Overview

<!-- 
Please provide an explanation of the PR, including the appropriate
context,
background, goal, and rationale. If there is an issue with this
information,
please provide a tl;dr and link the issue. 
-->

Context #1412 
Closes #1412 

Tested by running the test suite in a loop until the runner timeout with
no panics for writing to closed files:
https://github.com/rollkit/rollkit/actions/runs/7426536284/job/20210452949?pr=1434

## Checklist

<!-- 
Please complete the checklist to ensure that the PR is ready to be
reviewed.

IMPORTANT:
PRs should be left in Draft until the below checklist is completed.
-->

- [x] New and updated code has appropriate documentation
- [ ] New and updated code has new and/or updated testing
- [ ] Required CI checks are passing
- [ ] Visual proof for any user facing features like CLI or
documentation updates
- [x] Linked issues closed with keywords


<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->

## Summary by CodeRabbit

- **New Features**
- Implemented a new thread management system to enhance app performance
and stability.

- **Refactor**
- Improved internal thread handling mechanisms for better resource
management.

- **Tests**
- Added tests for the new thread management features to ensure
reliability.

<!-- end of auto-generated comment: release notes by coderabbit.ai -->