fix WebSocket serialization (#1425)
Unit test was updated to assert valid encoding of data. Hacky approach
to fix the issue is not working. Replacing normal JSON serialization
with JSON RPC serialization is required, but there still is an issue.

<!--
Please read and fill out this form before submitting your PR.

Please make sure you have reviewed our contributors guide before
submitting your
first PR.
-->

## Overview

Draft for visibility. CC: @Manav-Aggarwal 

Relates to: #1268 

<!-- 
Please provide an explanation of the PR, including the appropriate
context,
background, goal, and rationale. If there is an issue with this
information,
please provide a tl;dr and link the issue. 
-->

## Checklist

<!-- 
Please complete the checklist to ensure that the PR is ready to be
reviewed.

IMPORTANT:
PRs should be left in Draft until the below checklist is completed.
-->

- [ ] New and updated code has appropriate documentation
- [ ] New and updated code has new and/or updated testing
- [ ] Required CI checks are passing
- [ ] Visual proof for any user facing features like CLI or
documentation updates
- [ ] Linked issues closed with keywords


<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->

## Summary by CodeRabbit

- **New Features**
	- Enhanced WebSocket connection handling in JSON-RPC services.
  
- **Bug Fixes**
	- Improved subscription response encoding in JSON-RPC.

- **Refactor**
	- Updated WebSocket-related code to improve RPC request management.

- **Tests**
	- Altered test cases to reflect new JSON unmarshalling approach.

<!-- end of auto-generated comment: release notes by coderabbit.ai -->

---------

Co-authored-by: Ganesha Upadhyaya <ganeshrvce@gmail.com>
Co-authored-by: Ganesha Upadhyaya <gupadhyaya@Ganeshas-MacBook-Pro-2.local>