ignore validation failed error for the genesis block (#1499)
Fixes #969 

This PR fixes the `validation failed` error that is thrown by header and
block sync, which happens when the sequencer calls `Broadcast` which
also does publish the header/block locally for updating the store. This
happens because the sequencer before broadcasting, for the genesis
header, already does the store initialization and double store update
causes this error. The prior initialization is required to start the
syncer and there is no easier way to skip one of the two
initializations. In future, can refactor the code to move syncer start
to its own thread. But for now, ignoring the error is easier solution
without adding any safety concerns.

<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->
## Summary by CodeRabbit

- **Refactor**
- Improved handling of genesis blocks and headers during
synchronization, enhancing the app's stability and reducing unnecessary
error logs.
<!-- end of auto-generated comment: release notes by coderabbit.ai -->

---------

Co-authored-by: Ganesha Upadhyaya <gupadhyaya@Ganeshas-MacBook-Pro-2.local>