IBC for rollkit (JTMB IBC) (#1424)
<!--
Please read and fill out this form before submitting your PR.

Please make sure you have reviewed our contributors guide before
submitting your
first PR.
-->
ref: https://github.com/rollkit/rollkit/issues/1319
## Overview
Looking at the code, we found these issues when mapping a rollkit header
to an ABCI header:
- The signing byte for a rollkit header is different than that of an
ABCI header. This leads to the signature being invalid for the mapped
ABCI header.
- When mapping, we don't fill out some importants field in ABCI header
that is needed in the verifying process.

<!-- 
Please provide an explanation of the PR, including the appropriate
context,
background, goal, and rationale. If there is an issue with this
information,
please provide a tl;dr and link the issue. 
-->
## Changes 
- We temporarily add ValidatorHash to rollkit header so that the
ValidatorHash can be passed into the corresponding abci header. We're
finding a way to remove ValidatorHash and still make it work.
- Change in commit logic: senquencer will sign the `consensusVote`
instead of just the `header`.
- Make change when converting the rollkit commit to the ABCI commit for
compatibility with the tendermint light client.

### Testing
`TestRollKitABCIHeaderVerifiability` to ensure light client
compatibility, particularly to ensure that the rollkit header could be
verified by counterparty chain light client.
## Checklist

<!-- 
Please complete the checklist to ensure that the PR is ready to be
reviewed.

IMPORTANT:
PRs should be left in Draft until the below checklist is completed.
-->

- [ ] New and updated code has appropriate documentation
- [ ] New and updated code has new and/or updated testing
- [ ] Required CI checks are passing
- [ ] Visual proof for any user facing features like CLI or
documentation updates
- [ ] Linked issues closed with keywords


<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->
## Summary by CodeRabbit

- **New Features**
- Enhanced compatibility with Tendermint light client in block
management.
		- Improved block signing process.
- Updated block commitment handling to include app hash and validator
hash updates.
<!-- end of auto-generated comment: release notes by coderabbit.ai -->

---------

Co-authored-by: vuong177 <--global>
Co-authored-by: Hieu Vu <gundamaster5@gmail.com>
Co-authored-by: khanh <50263489+catShaark@users.noreply.github.com>
Co-authored-by: Damian Nolan <damiannolan@gmail.com>
Co-authored-by: Aditya Sripal <adityasripal@gmail.com>
Co-authored-by: Matthew Sevey <mjsevey@gmail.com>
Co-authored-by: Connor O'Hara <connor@switchboard.xyz>
Co-authored-by: s1nus <chuckhead7@gmail.com>
Co-authored-by: Manav Aggarwal <manavaggarwal1234@gmail.com>