Disable ISRs in Block Data (#1532)
<!--
Please read and fill out this form before submitting your PR.

Please make sure you have reviewed our contributors guide before
submitting your
first PR.
-->

## Overview


<!-- 
Please provide an explanation of the PR, including the appropriate
context,
background, goal, and rationale. If there is an issue with this
information,
please provide a tl;dr and link the issue. 
-->

Closes: https://github.com/rollkit/rollkit/issues/1531


## Checklist

<!-- 
Please complete the checklist to ensure that the PR is ready to be
reviewed.

IMPORTANT:
PRs should be left in Draft until the below checklist is completed.
-->

- [ ] New and updated code has appropriate documentation
- [ ] New and updated code has new and/or updated testing
- [x] Required CI checks are passing
- [ ] Visual proof for any user facing features like CLI or
documentation updates
- [x] Linked issues closed with keywords


<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->

## Summary by CodeRabbit

- **Refactor**
- Temporarily disabled the handling of `IntermediateStateRoots` across
various components of the system.
- **Tests**
	- Adjusted tests to skip scenarios involving `IntermediateStateRoots`.
- **Documentation**
- Updated internal documentation to reflect temporary changes in data
structure handling.

<!-- end of auto-generated comment: release notes by coderabbit.ai -->