Rearrange shutdown sequence and adjust clean-up in test
The shutdown sequence in the full node was rearranged to ensure proper closure of all services. Instead of immediately cancelling all node tasks and waiting for their termination, the update takes care of shutting down all sub services first before invoking cancellation and thread management wait. Furthermore, the clean-up process in the full node integration test was adjusted to delay node stoppage until after the state testing is done. This is to ensure that the store remains accessible for testing purposes.