Add Rollkit CLI (#791)
Closes: #745

## Key Update

Most notably this PR adds a `rollkit` CLI tool. Here are some of the
changes that come along with it.

### CLI Docs

A `docs-gen` command was added to the CLI to auto generate the
documentation for the CLI tool. This documentation likes in
`./cmd/rollkit/docs`.

These Markdown files are then ignore by the markdown linter.

### Makefile changes

Currently rollkit is used via the cosmos-sdk. The outdated reference to
`go build` in the README was removed as there are now current requests
for using a rollkit binary.

This allows for the default `make install` to be for the rollkit CLI. 

### Commands

The initial commands of the CLI are:
- `docs-gen`
- `start`
- `version`

`docs-gen` has been previously described. 

`rollkit start` is the primary command. It handles initialization as a
pre-run step and launches the node for the user.

`version` is used to see the current version that the rollkit CLI is
using. This is dynamically set from the git history.


----------



<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->
## Summary by CodeRabbit

- **New Features**
- Introduced a modular framework for rollups with a comprehensive CLI
for configuration, logging, and error handling.
- Added autocompletion script generation commands for bash, fish,
PowerShell, and zsh shells.
- Implemented a `rollkit start` command to run the rollkit node with
extensive configuration options.
- Provided a `rollkit version` command to display version information
for Rollkit.
- Introduced documentation generation functionality for the rollkit CLI.
- **Documentation**
- Added detailed guides for setting up autocompletion in various shells.
- Introduced documentation on running the rollkit node and displaying
version information.
- **Chores**
- Implemented a new lint job for CLI documentation to ensure quality and
consistency.
<!-- end of auto-generated comment: release notes by coderabbit.ai -->

---------

Co-authored-by: Matthew Sevey <mjsevey@gmail.com>
Co-authored-by: Ganesha Upadhyaya <ganeshrvce@gmail.com>
Co-authored-by: Ganesha Upadhyaya <gupadhyaya@Ganeshas-MacBook-Pro-2.local>