test : Use IsType assertions in test (#1574)
<!--
Please read and fill out this form before submitting your PR.

Please make sure you have reviewed our contributors guide before
submitting your
first PR.
-->

## Overview
closes : #1570 
<!-- 
Please provide an explanation of the PR, including the appropriate
context,
background, goal, and rationale. If there is an issue with this
information,
please provide a tl;dr and link the issue. 
-->

## Checklist

<!-- 
Please complete the checklist to ensure that the PR is ready to be
reviewed.

IMPORTANT:
PRs should be left in Draft until the below checklist is completed.
-->

- [ ] New and updated code has appropriate documentation
- [ ] New and updated code has new and/or updated testing
- [ ] Required CI checks are passing
- [ ] Visual proof for any user facing features like CLI or
documentation updates
- [ ] Linked issues closed with keywords


<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->
## Summary by CodeRabbit


- **Tests**
- Enhanced testing for both Full and Light Node types to ensure correct
object types are created and utilized.
- Added assertions using `require.IsType` in various test functions to
check object types.
- Added assertion in `TestStartup` and `TestMempoolDirectly` to validate
the type of the node object as `FullNode`.
- Introduced a new assertion in `TestLightClient_Panics` to verify the
type of `ln`.
- Updated `TestNewNode` to include assertions using `require.IsType` to
validate the types of nodes created, specifically for `LightNode` and
`FullNode`.

<!-- end of auto-generated comment: release notes by coderabbit.ai -->