Refactor GetRandomSignedHeaders and GetValidatorSet Helpers (#1564)
Cleaning up helper methods to adhere to a common pattern of:
```golang
GetXYZ()
GetXYZCustom()
```

Previously, we were heading down a not very DRY path of GetXYZWithABC
where ABC was just some config change.
This PR DRYs up these helper methods so that only the Custom methods
need to be extended in the future, and the majority of tests can
continue to use the basic Get method without
issue.
<!--
Please read and fill out this form before submitting your PR.

Please make sure you have reviewed our contributors guide before
submitting your
first PR.
-->

## Overview

<!-- 
Please provide an explanation of the PR, including the appropriate
context,
background, goal, and rationale. If there is an issue with this
information,
please provide a tl;dr and link the issue. 
-->

## Checklist

<!-- 
Please complete the checklist to ensure that the PR is ready to be
reviewed.

IMPORTANT:
PRs should be left in Draft until the below checklist is completed.
-->

- [ ] New and updated code has appropriate documentation
- [ ] New and updated code has new and/or updated testing
- [ ] Required CI checks are passing
- [ ] Visual proof for any user facing features like CLI or
documentation updates
- [ ] Linked issues closed with keywords


<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->
## Summary by CodeRabbit

- **Refactor**
- Enhanced the flexibility and customization in block and validator set
generation for testing purposes.
- **Tests**
- Improved test configurations for more accurate simulation of various
scenarios.
- **Chores**
- Renamed several core functions and structs to better reflect their
purpose and usage.
<!-- end of auto-generated comment: release notes by coderabbit.ai -->

---------

Co-authored-by: Manav Aggarwal <manavaggarwal1234@gmail.com>