Refactor pending blocks handling (#1568)
<!--
Please read and fill out this form before submitting your PR.

Please make sure you have reviewed our contributors guide before
submitting your
first PR.
-->

## Overview

This PR creates a new test to check pending blocks as described in issue
#1548. It mimics the behavior of a node producing blocks, stopping and
restarting, then producing more blocks.

Looking at commit history you can check that test was initially failing
and then was fixed.

New implementation of `pendingBlocks` is safer because its data is
persisted in store. If node is restarted, block submission to DA will
restart when it ended. The worst case scenario (very unlikely), where
`pendingBlocks` can't save information to store, results in resubmission
of blocks already submitted to DA, which is only the extra cost. The
solution ensures that all the blocks are successfully submitted to DA.

This PR is bigger than I expected, and because of that I will work on
https://github.com/rollkit/rollkit/issues/1524 and limiting the number
of blocks returned by `getPendingBlocks` in follow-up PR.

Resolves #1548
Resolves #457 

<!-- 
Please provide an explanation of the PR, including the appropriate
context,
background, goal, and rationale. If there is an issue with this
information,
please provide a tl;dr and link the issue. 
-->

## Checklist

<!-- 
Please complete the checklist to ensure that the PR is ready to be
reviewed.

IMPORTANT:
PRs should be left in Draft until the below checklist is completed.
-->

- [x] New and updated code has appropriate documentation
- [x] New and updated code has new and/or updated testing
- [x] Required CI checks are passing
- [x] Visual proof for any user facing features like CLI or
documentation updates
- [x] Linked issues closed with keywords


<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->
## Summary by CodeRabbit

- **New Features**
- Introduced a new mock generation command for enhanced testing
capabilities.
- Added functionality for improved block management and submission to
Data Availability layers.
- Enhanced metadata management in the store for better data handling and
retrieval.
- Implemented new testing functions and improved existing ones for
better coverage and reliability.

- **Bug Fixes**
- Fixed handling of empty block submissions to prevent errors during
block publishing.

- **Refactor**
- Refactored block management to improve code efficiency and
readability.
- Consolidated mock application creation logic in integration tests for
better maintainability.

- **Tests**
- Expanded testing suite with new tests for block submission, metadata
operations, and mock DA interactions.

- **Chores**
- Updated dependencies and imports across various files to enhance
functionality and testing.
<!-- end of auto-generated comment: release notes by coderabbit.ai -->