fix(manager): reduce the number of isProposer checks in manager (#1555)
<!--
Please read and fill out this form before submitting your PR.

Please make sure you have reviewed our contributors guide before
submitting your
first PR.
-->

## Overview

Previously, IsProposal() was called to check if the validator key from
Genesis data matches the proposer's public key every time `publishBlock`
is called. Although cheap to compute, this call was redundant. This PR
has simple changes to cache it at the manager struct level during
instantiation.

<!-- 
Please provide an explanation of the PR, including the appropriate
context,
background, goal, and rationale. If there is an issue with this
information,
please provide a tl;dr and link the issue. 
-->

## Checklist

<!-- 
Please complete the checklist to ensure that the PR is ready to be
reviewed.

IMPORTANT:
PRs should be left in Draft until the below checklist is completed.
-->

Closes https://github.com/rollkit/rollkit/issues/1452

- [ ] New and updated code has appropriate documentation
- [ ] New and updated code has new and/or updated testing
- [ ] Required CI checks are passing
- [ ] Visual proof for any user facing features like CLI or
documentation updates
- [ ] Linked issues closed with keywords


<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->
## Summary by CodeRabbit


- **Refactor**
- Improved block publishing logic by utilizing a new internal flag to
determine if the current node is the proposer.
- **Tests**
- Added tests to validate proposer determination and block publishing
behavior for both proposer and non-proposer scenarios.

<!-- end of auto-generated comment: release notes by coderabbit.ai -->