ci: add ci for the cli for rollkit start cmd (#1603)
<!--
Please read and fill out this form before submitting your PR.

Please make sure you have reviewed our contributors guide before
submitting your
first PR.
-->

## Overview
We will want to add this new CLI CI to the required checks when this
gets merged in.
As such, I won't use auto merge on this so that I can update the branch
protection rules right after I merge this.
<!-- 
Please provide an explanation of the PR, including the appropriate
context,
background, goal, and rationale. If there is an issue with this
information,
please provide a tl;dr and link the issue. 
-->

## Checklist

<!-- 
Please complete the checklist to ensure that the PR is ready to be
reviewed.

IMPORTANT:
PRs should be left in Draft until the below checklist is completed.
-->

- [ ] New and updated code has appropriate documentation
- [ ] New and updated code has new and/or updated testing
- [ ] Required CI checks are passing
- [ ] Visual proof for any user facing features like CLI or
documentation updates
- [ ] Linked issues closed with keywords


<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->
## Summary by CodeRabbit


- **New Features**
- Introduced a new job in the CI workflow specifically for testing the
CLI with enhanced parameters.
- **Refactor**
- Updated the CLI's run node command to support CI mode, preventing
indefinite blocking during automated tests by adding a new `ci` flag.

<!-- end of auto-generated comment: release notes by coderabbit.ai -->