fix(manager): refactor lazy mode flags (#1800)
## Overview

Following up #1779 the new flag `LazyBufferTime` has been removed to
simplify lazy mode flags. We use a hardcoded value of
`defaultLazySleepPercent = 10` to indicate a wait time equal to 10% of
the block time. It might be possible to expose this as a flag if users
request.


<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->
## Summary by CodeRabbit


- **New Features**
- Enhanced configurability of transaction accumulation by shifting from
a static time-based approach to a dynamic percentage-based system.
	
- **Bug Fixes**
- Improved accuracy in sleep duration calculations related to block
manager behavior.

- **Chores**
- Simplified node configuration by removing the `LazyBufferTime`
parameter, streamlining operational parameters.
- Simplified command-line options by removing the
`--rollkit.lazy_buffer_time` parameter.

<!-- end of auto-generated comment: release notes by coderabbit.ai -->