feat: create core pkg  (#2012)
This is the base pr for combining the sequencing and execution systems
into a single.


the goal of this design is to move rollkit into a library and less so a
binary. By default there will be a centralized and based sequencer
binaries with abci and reth. These will serve as examples as well to how
to compose your own app.

as part of this i would propose we mono repo as we move things around. 

we would create a sequencer folder with each sequencer design having its
own go.mod.

execution could live in a different repo for now, for the simple ones we
can have an execution folder with each execution env having its own
go.mod but for things more complex like abci it would probably be good
to have it in its own repo.

---------

Co-authored-by: tac0turtle <markobaricevic@MacBookPro.localdomain>