feat: add load from viper (#2176)
Ref:
https://github.com/rollkit/go-execution-abci/pull/48#discussion_r2044292045
I eventually went with `LoadFromViper`, as the flags are added in the
start command and i would have had to duplicate most of the logic
otherwise.

`LoadFromViper` should be used for the SDK, as it does some
prepossessing of flags. `Load` should be used for any other use case
(non SDK based chains).