fix: return soft-confirmed batches (#2180)
## Overview

Return batches before DA submission, to make Single sequencer compatible
to the spec (ADR).

Resolves #2179


<!-- 
Please provide an explanation of the PR, including the appropriate
context,
background, goal, and rationale. If there is an issue with this
information,
please provide a tl;dr and link the issue. 

Ex: Closes #<issue number>
-->